refactor(macro_common): split concept test out
