+++ shared_crates/project_macro_common/src/tests.rs
@@ -1,13 +1,9 @@
-use std::collections::BTreeMap;
-
-use super::{concept::generate_component_list_doc_comment, implementation, tree::Tree, Context};
-use ambient_project::{Component, ComponentType, Concept, IdentifierPathBuf};
+use super::{implementation, Context};
 use proc_macro2::Span;
 
-fn context() -> Context {
-    let ident = syn::Ident::new("ambient_api2", Span::call_site());
+pub(crate) fn guest_context() -> Context {
     Context::Guest {
-        api_path: ident.into(),
+        api_path: syn::Ident::new("ambient_api2", Span::call_site()).into(),
         fully_qualified_path: true,
     }
 }
@@ -96,7 +92,7 @@ fn can_generate_components_from_manifest_in_global_namespace() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         true,
         true,
     )
@@ -135,7 +131,7 @@ fn can_accept_no_components() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     )
@@ -196,7 +192,7 @@ fn can_generate_components_from_manifest() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     )
@@ -244,7 +240,7 @@ fn can_generate_component_with_contained_type() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     )
@@ -293,7 +289,7 @@ fn can_generate_components_from_manifest_with_org() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     )
@@ -346,7 +342,7 @@ fn can_generate_components_with_documented_namespace_from_manifest() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     )
@@ -369,7 +365,7 @@ fn will_error_on_undocumented_namespace() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     );
@@ -570,7 +566,7 @@ fn can_generate_concepts_with_all_supported_types() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         false,
     )
@@ -676,7 +672,7 @@ fn can_extend_with_multiple_concepts() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         false,
     )
@@ -844,7 +840,7 @@ fn can_generate_concepts() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         false,
     )
@@ -948,142 +944,13 @@ fn can_generate_concepts_with_documented_namespace_from_manifest() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         false,
     )
     .unwrap();
 
     assert_eq!(result.to_string(), expected_output.to_string());
-}
-
-#[test]
-fn can_generate_nested_doc_comment_for_concepts() {
-    let component_tree = Tree::<Component>::new(
-        &BTreeMap::from_iter(
-            ["Bool", "F32", "I32", "String", "Vec3"]
-                .into_iter()
-                .enumerate()
-                .map(|(idx, ty)| {
-                    (
-                        IdentifierPathBuf::new(format!("component{idx}")).unwrap(),
-                        Component {
-                            name: format!("Component {idx}"),
-                            description: "".to_string(),
-                            type_: ComponentType::String(ty.to_string()),
-                            attributes: vec![],
-                            default: None,
-                        }
-                        .into(),
-                    )
-                }),
-        ),
-        false,
-    )
-    .unwrap();
-
-    let concept_tree = Tree::<Concept>::new(
-        &BTreeMap::from_iter([
-            (
-                IdentifierPathBuf::new("concept0").unwrap(),
-                Concept {
-                    name: String::new(),
-                    description: String::new(),
-                    extends: vec![],
-                    components: BTreeMap::from_iter([(
-                        IdentifierPathBuf::new("component0").unwrap(),
-                        toml::Value::Boolean(true),
-                    )]),
-                }
-                .into(),
-            ),
-            (
-                IdentifierPathBuf::new("concept1").unwrap(),
-                Concept {
-                    name: String::new(),
-                    description: String::new(),
-                    extends: vec![IdentifierPathBuf::new("concept0").unwrap()],
-                    components: BTreeMap::from_iter([(
-                        IdentifierPathBuf::new("component1").unwrap(),
-                        toml::Value::Float(3.14),
-                    )]),
-                }
-                .into(),
-            ),
-            (
-                IdentifierPathBuf::new("concept2").unwrap(),
-                Concept {
-                    name: String::new(),
-                    description: String::new(),
-                    extends: vec![],
-                    components: BTreeMap::from_iter([(
-                        IdentifierPathBuf::new("component2").unwrap(),
-                        toml::Value::Integer(3),
-                    )]),
-                }
-                .into(),
-            ),
-            (
-                IdentifierPathBuf::new("concept3").unwrap(),
-                Concept {
-                    name: String::new(),
-                    description: String::new(),
-                    extends: vec![
-                        IdentifierPathBuf::new("concept1").unwrap(),
-                        IdentifierPathBuf::new("concept2").unwrap(),
-                    ],
-                    components: BTreeMap::from_iter([
-                        (
-                            IdentifierPathBuf::new("component3").unwrap(),
-                            toml::Value::String("It's pi".to_string()),
-                        ),
-                        (
-                            IdentifierPathBuf::new("component4").unwrap(),
-                            toml::Value::Array(
-                                (0..3).into_iter().map(toml::Value::Integer).collect(),
-                            ),
-                        ),
-                    ]),
-                }
-                .into(),
-            ),
-        ]),
-        false,
-    )
-    .unwrap();
-
-    let comment = generate_component_list_doc_comment(
-        &concept_tree,
-        &component_tree,
-        &context(),
-        concept_tree
-            .get(IdentifierPathBuf::new("concept3").unwrap().as_path())
-            .unwrap(),
-    )
-    .unwrap();
-
-    assert_eq!(
-        comment,
-        indoc::indoc! {r#"
-            *Definition*:
-
-            ```
-            {
-              "component3": String = "It's pi".to_string(),
-              "component4": Vec3 = Vec3::new(0f32, 1f32, 2f32),
-              "concept1": { // Concept.
-                "component1": f32 = 3.14f32,
-                "concept0": { // Concept.
-                  "component0": bool = true,
-                },
-              },
-              "concept2": { // Concept.
-                "component2": i32 = 3i32,
-              },
-            }
-            ```
-        "#}
-    );
 }
 
 #[test]
@@ -1159,7 +1026,7 @@ fn can_generate_message() {
 
     let result = implementation(
         (Some("ambient.toml".to_string()), manifest.to_string()),
-        context(),
+        guest_context(),
         false,
         true,
     )
